+++ tests/providers/azure/services/databricks/databricks_service_test.py
@@ -3,6 +3,7 @@ from unittest.mock import patch
 from prowler.providers.azure.services.databricks.databricks_service import (
     Databricks,
     DatabricksWorkspace,
+    ManagedDiskEncryption,
 )
 from tests.providers.azure.azure_fixtures import (
     AZURE_SUBSCRIPTION_ID,
@@ -18,6 +19,11 @@ def mock_databricks_get_workspaces(_):
                 name="test-workspace",
                 location="eastus",
                 custom_managed_vnet_id="test-vnet-id",
+                managed_disk_encryption=ManagedDiskEncryption(
+                    key_name="test-key",
+                    key_version="test-version",
+                    key_vault_uri="test-vault-uri",
+                ),
             )
         }
     }
@@ -48,3 +54,39 @@ class Test_Databricks_Service:
         assert workspace.name == "test-workspace"
         assert workspace.location == "eastus"
         assert workspace.custom_managed_vnet_id == "test-vnet-id"
+        assert (
+            workspace.managed_disk_encryption.__class__.__name__
+            == "ManagedDiskEncryption"
+        )
+        assert workspace.managed_disk_encryption.key_name == "test-key"
+        assert workspace.managed_disk_encryption.key_version == "test-version"
+        assert workspace.managed_disk_encryption.key_vault_uri == "test-vault-uri"
+
+
+def mock_databricks_get_workspaces_no_encryption(_):
+    return {
+        AZURE_SUBSCRIPTION_ID: {
+            "test-workspace-id": DatabricksWorkspace(
+                id="test-workspace-id",
+                name="test-workspace",
+                location="eastus",
+                custom_managed_vnet_id="test-vnet-id",
+                managed_disk_encryption=None,
+            )
+        }
+    }
+
+
+@patch(
+    "prowler.providers.azure.services.databricks.databricks_service.Databricks._get_workspaces",
+    new=mock_databricks_get_workspaces_no_encryption,
+)
+class Test_Databricks_Service_No_Encryption:
+    def test_get_workspaces_no_encryption(self):
+        databricks = Databricks(set_mocked_azure_provider())
+        workspace = databricks.workspaces[AZURE_SUBSCRIPTION_ID]["test-workspace-id"]
+        assert workspace.id == "test-workspace-id"
+        assert workspace.name == "test-workspace"
+        assert workspace.location == "eastus"
+        assert workspace.custom_managed_vnet_id == "test-vnet-id"
+        assert workspace.managed_disk_encryption is None